app/gateway/pyproject.toml
[project]
name = "gnosis-gateway"
version = "0.1.0"
description = "Gnosis API Gateway"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "pydantic",
    "fastapi",
    "uvicorn",
    "grpcio",
    "grpcio-tools",
    "modal",
    "opencv-python",
    "python-dotenv",
    "requests",
    "Pillow",
    "python-multipart",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = [
  "data",
  "build/",
  "dist/",
  ".venv/",
  "__pycache__",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
# which rule sets to enable; you can start simple
select = ["E", "F"]  # errors (E) + pyflakes (F)
ignore = []           # none to ignore initially
